requirements.txt
Django	1.10.4
Pillow	4.0.0
django-address	0.1.5
django-filter	1.0.1
djangorestframework	3.5.3